Games/quizshow/requirements.txt
sqlalchemy
flask
pyserial
requests
pathlib
toml
websocket-client